slide update
