+++ src/styles/TripCourse.css
@@ -1,5 +1,4 @@
 
-
 .slider {
   position: relative;
   max-width: 100%;
@@ -9,16 +8,16 @@
   margin-top: 200px;
 }
 
-.sliderback {
+.tripCourse-slider-background {
   width: 100%;
   height: 400px;
   background-color: #8CA1EB;
   z-index: 0;
   margin-top: 100px;
-  
+
 }
 
-.sliderfont {
+.tripCourse-slider-title {
   width: 600px;
   height: 100px;
   margin-left: 760px;
@@ -26,18 +25,16 @@
   display: inline-block;
   position: relative; 
   font-weight: bold;
-
-  
 }
 
-.sliderfont::after {
+.tripCourse-slider-title::after {
   content: '';
   display: block;
   width: 100%;
   height: 3px;
   background-color: black;
   position: absolute;
-  bottom: 30px; 
+   top: 50px;
   left: 0;
 }
 
@@ -53,15 +50,20 @@
   box-shadow: 10px 10px 20px rgba(0, 0, 0, 0.5);
 }
 
-.sliderread {
+.tripCourse-slider-contents-list {
   width: 600px;
   height: 200px;
   text-align: left;
   position: relative;
- font-size: 30px;
-  margin-left: 760px; 
+  bottom: 80px;
+  font-size: 25px;
+  margin-left: 760px;
 }
 
+.slider img {
+  width: 600px;
+  height: 350px;
+}
 
 
 .lodging {
